pyproject.toml
[project]
name = "hive-mcp-server"
version = "0.1.0"
description = "Cloudera Hive MCP Server via Impala"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.9.2",
    "impyla>=0.21.0",
    "mcp[cli]>=1.9.4",
    "python-dotenv>=1.1.0",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.1.0", 
    "openai>=1.0.0",
]

[dependency-groups]
dev = [
    "autoflake==2.3.1",
    "debugpy==1.8.12",
    "pytest==8.2.2",
    "pytest-cov==5.0.0",
    "ruff==0.11.10",
]

[tool.ruff]
line-length = 120

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/hive_mcp_server/__init__.py"

[project.scripts]
run-server = "hive_mcp_server.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hive_mcp_server"]
